rename files, use makefile to set order
