+++ high_quality_transit_areas/Makefile
@@ -1,12 +1,12 @@
 hqta_data:
-	python A1_rail_ferry_brt_stops.py
-	python B1_create_hqta_segments.py
-	python B2_sjoin_stops_to_segments.py
-	python C1_prep_pairwise_intersections.py
-	python C2_get_intersections.py
-	python C3_create_bus_hqta_types.py
-	python D1_assemble_hqta_points.py
-	python D2_assemble_hqta_polygons.py 
+	python rail_ferry_brt_stops.py
+	python create_hqta_segments.py
+	python sjoin_stops_to_segments.py
+	python prep_pairwise_intersections.py
+	python get_intersections.py
+	python create_bus_hqta_types.py
+	python assemble_hqta_points.py
+	python assemble_hqta_polygons.py 
 
 # Only need this is operator input changes
 # For now, Muni sent over a date-versioned list of stops
