+++ high_quality_transit_areas/catalog.yml
@@ -14,69 +14,69 @@ sources:
         args:
           urlpath: gs://calitp-analytics-data/data-analyses/high_quality_transit_areas/ca_boundary.parquet
     ## INTERMEDIATE DATA
-    # Source: A1_rail_ferry_brt_stops.py
+    # Source: rail_ferry_brt_stops.py
     stops_info_crosswalk:
         driver: geoparquet
         description: Assembled stop gdf with route info from trips table.
         args:
           urlpath: gs://calitp-analytics-data/data-analyses/high_quality_transit_areas/stops_to_route.parquet
-    # Source: A1_rail_ferry_brt_stops.py
-    rail_brt_ferry_initial:
+    # Source: rail_ferry_brt_stops.py
+    rail_brt_ferry_stops:
         driver: geoparquet
         description: All the major transit stops (Rail / BRT / Ferry).
         args:
           urlpath: gs://calitp-analytics-data/data-analyses/high_quality_transit_areas/rail_brt_ferry.parquet
-    # Source: B1_create_hqta_segments.py
+    # Source: create_hqta_segments.py
     hqta_segments:
         driver: geoparquet
-        description: Cut HQTA segments across all operators. Created in B1_create_hqta_segments.py.
+        description: Cut HQTA segments across all operators. 
         args:
           urlpath: gs://calitp-analytics-data/data-analyses/high_quality_transit_areas/hqta_segments.parquet   
-    # Source: B2_sjoin_stops_to_segments.py
+    # Source: sjoin_stops_to_segments.py
     all_bus:
         driver: geoparquet
-        description: Combined hqta corridors across all operators. Created in B2_sjoin_stops_to_segments.py.
+        description: Combined hqta corridors across all operators. 
         args:
           urlpath: gs://calitp-analytics-data/data-analyses/high_quality_transit_areas/all_bus.parquet  
-    # Source: C1_prep_pairwise_intersections.py
+    # Source: prep_pairwise_intersections.py
     pairwise_intersections:
         driver: parquet
-        description: Use spatial join to find which hqta segments do intersect at some point. Created in C1_prep_pairwise_intersections.py. 
+        description: Use spatial join to find which hqta segments do intersect at some point.
         args:
           urlpath: gs://calitp-analytics-data/data-analyses/high_quality_transit_areas/pairwise.parquet   
-    # Source: C1_prep_pairwise_intersections.py
+    # Source: prep_pairwise_intersections.py
     subset_corridors:
         driver: geoparquet
-        description: Narrow down `all_bus` to hqta segments that are found in `pairwise_intersections`. Created in C1_prep_pairwise_intersections.py.
+        description: Narrow down `all_bus` to hqta segments that are found in `pairwise_intersections`.
         args:
           urlpath: gs://calitp-analytics-data/data-analyses/high_quality_transit_areas/subset_corridors.parquet 
-    # Source: C2_get_intersections.py
+    # Source: get_intersections.py
     all_intersections:
         driver: geoparquet
         description: Find where 2 bus corridors intersect by doing an intersection between the hqta segments.
         args:
           urlpath: gs://calitp-analytics-data/data-analyses/high_quality_transit_areas/all_intersections.parquet
-    # Source: C3_create_bus_hqta_types.py
+    # Source: create_bus_hqta_types.py
     major_stop_bus:
         driver: geoparquet
-        description: Bus stops that are within the bus intersections. Created in C3_create_bus_hqta_types.py.
+        description: Bus stops that are within the bus intersections. 
         args:
           urlpath: gs://calitp-analytics-data/data-analyses/high_quality_transit_areas/major_stop_bus.parquet
     stops_in_hq_corr:
         driver: geoparquet
-        description: Bus stops that are within the HQ corridors, even if they stops themselves do not have a lot of trips pass through. Created in C3_create_bus_hqta_types.py.
+        description: Bus stops that are within the HQ corridors, even if they stops themselves do not have a lot of trips pass through. 
         args:
           urlpath: gs://calitp-analytics-data/data-analyses/high_quality_transit_areas/stops_in_hq_corr.parquet
     ## FINAL DATA
-    # Source: D1_assemble_hqta_points
+    # Source: assemble_hqta_points
     hqta_points:
         driver: geoparquet
-        description: Combined point data for all HQTA types. Created in D1_assemble_hqta_points.py.
+        description: Combined point data for all HQTA types.
         args:
           urlpath: gs://calitp-analytics-data/data-analyses/high_quality_transit_areas/hqta_points.parquet
-    # Source: D2_assemble_hqta_polygons
+    # Source: assemble_hqta_polygons
     hqta_areas:
         driver: geoparquet
-        description: Combined polygon data for all HQTA types. Created in D2_assemble_hqta_polygons.py.
+        description: Combined polygon data for all HQTA types. 
         args:
           urlpath: gs://calitp-analytics-data/data-analyses/high_quality_transit_areas/hqta_areas.parquet
